Fixed more warnings

+++ src/MainLoopDriver.cpp
@@ -96,7 +96,7 @@ int main()
 
 
     // Distribute territories
-    for (int i = 0; i < gameEngine.getMap().territories.size(); ++i)
+    for (size_t i = 0; i < gameEngine.getMap().territories.size(); ++i)
     {
         Player* playerToAssign = gameEngine.getPlayers().at(i % numPlayers);
